requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
Flask==1.1.2
openpyxl==3.0.3
Pillow==7.1.2
pytesseract==0.3.4
selenium==3.141.0
tesseract==0.1.3